Added argument to list the registered features Moved some scripts around to clean up the hierarchy
