+++ documentation/propertiesconfig.md
